requirement.txt
dash
pillow
dash-daq
plotly>=5.3.1
